src/entrypoint_service/requirements.txt
flask~=3.0.2
requests~=2.31.0
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
